Adopt the fix
