+++ client/tso_dispatcher.go
@@ -357,14 +357,22 @@ func (c *tsoClient) handleDispatcher(
 	if dc == globalDCLocation {
 		go func() {
 			var updateTicker = &time.Ticker{}
-			setNewUpdateTicker := func(ticker *time.Ticker) {
+			setNewUpdateTicker := func(interval time.Duration) {
 				if updateTicker.C != nil {
 					updateTicker.Stop()
 				}
-				updateTicker = ticker
+				if interval == 0 {
+					updateTicker = &time.Ticker{}
+				} else {
+					updateTicker = time.NewTicker(interval)
+				}
+			}
+			// If the TSO Follower Proxy is enabled, set the update interval to the member update interval.
+			if c.option.getEnableTSOFollowerProxy() {
+				setNewUpdateTicker(memberUpdateInterval)
 			}
 			// Set to nil before returning to ensure that the existing ticker can be GC.
-			defer setNewUpdateTicker(nil)
+			defer setNewUpdateTicker(0)
 
 			for {
 				select {
@@ -375,14 +383,11 @@ func (c *tsoClient) handleDispatcher(
 					if enableTSOFollowerProxy && updateTicker.C == nil {
 						// Because the TSO Follower Proxy is enabled,
 						// the periodic check needs to be performed.
-						setNewUpdateTicker(time.NewTicker(memberUpdateInterval))
+						setNewUpdateTicker(memberUpdateInterval)
 					} else if !enableTSOFollowerProxy && updateTicker.C != nil {
 						// Because the TSO Follower Proxy is disabled,
 						// the periodic check needs to be turned off.
-						setNewUpdateTicker(&time.Ticker{})
-					} else {
-						// The status of TSO Follower Proxy does not change, and updateTSOConnectionCtxs is not triggered
-						continue
+						setNewUpdateTicker(0)
 					}
 				case <-updateTicker.C:
 				case <-c.updateTSOConnectionCtxsCh:
